+++ syscore/dateutils.py
@@ -8,6 +8,7 @@ import datetime
 import time
 import calendar
 import pandas as pd
+import numpy as np
 
 from syscore.objects import missing_data, arg_not_supplied
 
@@ -43,6 +44,73 @@ UNIXTIME_CONVERTER = 1e9
 UNIXTIME_IN_YEAR = UNIXTIME_CONVERTER * SECONDS_IN_YEAR
 
 MONTH_LIST = ["F", "G", "H", "J", "K", "M", "N", "Q", "U", "V", "X", "Z"]
+
+def get_date_from_period_and_end_date(period: str,
+                                      end_date = arg_not_supplied) -> datetime.datetime:
+    if end_date is arg_not_supplied:
+        end_date = datetime.datetime.now()
+
+    if period=="YTD":
+        return calculate_starting_day_of_current_year(end_date)
+    elif period=="MTD":
+        return calculate_starting_day_of_current_month(end_date)
+    elif period=="QTD":
+        return calculate_starting_day_of_current_quarter(end_date)
+    elif period=="TAX":
+        return calculate_starting_day_of_current_uk_tax_year(end_date)
+
+    offset_date =  _get_previous_date_from_period_with_char_number(period, end_date)
+    return offset_date
+
+def _get_previous_date_from_period_with_char_number(period: str,
+                          end_date: datetime.datetime) -> datetime.datetime:
+
+    type_of_offset = period[-1]
+    try:
+        number_offset = int(period[:1])
+    except:
+        raise Exception("Offset %s not in pattern numberLetter eg 7D for 7 days")
+
+    if type_of_offset=="M":
+        offset_date = end_date+pd.DateOffset(months=-number_offset)
+    elif type_of_offset=="D":
+        offset_date = end_date+pd.DateOffset(days = -number_offset)
+    elif type_of_offset=="W":
+        offset_date = end_date + pd.DateOffset(days=-number_offset*7)
+    elif type_of_offset=="B":
+        offset_date = end_date+pd.tseries.offsets.BDay(-number_offset)
+    elif type_of_offset=="Y":
+        offset_date = end_date+pd.DateOffset(years = -number_offset)
+    else:
+        raise Exception("Type of offset %s not recognised must be one of BDMYW" % type_of_offset)
+
+    return offset_date
+
+def calculate_starting_day_of_current_year(end_date) -> datetime.datetime:
+    return datetime.datetime(year=end_date.year,
+                             month=1,
+                             day=1)
+
+def calculate_starting_day_of_current_uk_tax_year(end_date: datetime.datetime) -> datetime.datetime:
+    current_month = end_date.month
+    current_day = end_date.day
+    current_year = end_date.year
+    if current_month<5 and current_day<6:
+        return datetime.datetime(year=current_year-1, month=4, day=6)
+    else:
+        return datetime.datetime(year = current_year, month=4, day=6)
+
+def calculate_starting_day_of_current_month(end_date) -> datetime.datetime:
+    return datetime.datetime(year = end_date.year, month = end_date.month, day=1)
+
+def calculate_starting_day_of_current_quarter(end_date):
+    current_month = end_date.month
+    current_quarter = int(np.ceil(current_month/3))
+    start_month_of_quarter = ((current_quarter -1)*3)+1
+    return datetime.datetime(year = end_date.year,
+                             month = start_month_of_quarter,
+                             day=1)
+
 
 
 Frequency = Enum(
@@ -291,42 +359,6 @@ def strip_timezone_fromdatetime(timestamp_with_tz_info) -> datetime.datetime:
     return new_timestamp
 
 
-def get_datetime_input(
-    prompt: str, allow_default: bool = True, allow_no_arg: bool = False
-):
-    invalid_input = True
-    input_str = (
-        prompt
-        + ": Enter date and time in format %Y-%m-%d eg '2020-05-30' OR '%Y-%m-%d %H:%M:%S' eg '2020-05-30 14:04:11'"
-    )
-    if allow_default:
-        input_str = input_str + " <RETURN for now>"
-    if allow_no_arg:
-        input_str = input_str + " <SPACE for no date>' "
-    while invalid_input:
-        ans = input(input_str)
-        if ans == "" and allow_default:
-            return datetime.datetime.now()
-        if ans == " " and allow_no_arg:
-            return None
-        try:
-            if len(ans) == 10:
-                return_datetime = datetime.datetime.strptime(ans, "%Y-%m-%d")
-            elif len(ans) == 19:
-                return_datetime = datetime.datetime.strptime(ans, "%Y-%m-%d %H:%M:%S")
-            else:
-                # problems formatting will also raise value error
-                raise ValueError
-            return return_datetime
-
-        except ValueError:
-            print("%s is not a valid datetime string" % ans)
-            continue
-
-
-
-
-
 SHORT_DATE_PATTERN = "%m/%d %H:%M:%S"
 MISSING_STRING_PATTERN = "     ???      "
 
@@ -536,3 +568,52 @@ def generate_equal_dates_within_year(
     ]
 
     return all_dates
+
+
+def get_approx_vol_scalar_for_period(start_date:datetime.datetime,
+                              end_date = datetime.datetime) -> float:
+    time_between = end_date - start_date
+    seconds_between= time_between.total_seconds()
+    days_between = seconds_between / SECONDS_PER_DAY
+
+    return days_between**.5
+
+
+def calculate_start_and_end_dates(
+        calendar_days_back = arg_not_supplied,
+        end_date: datetime.datetime = arg_not_supplied,
+        start_date: datetime.datetime = arg_not_supplied,
+        start_period: str = arg_not_supplied,
+        end_period: str = arg_not_supplied) -> tuple:
+
+    ## DO THE END DATE FIRST
+    ## First preference is to use period
+    if end_period is arg_not_supplied:
+        ## OK preference is to use passed date, if there was one
+        if end_date is arg_not_supplied:
+            end_date = datetime.datetime.now()
+        else:
+            # Use passed end date
+            pass
+    else:
+        end_date = get_date_from_period_and_end_date(end_period)
+
+    ## DO THE START DATE NEXT
+    ## First preference is to use period, then calendar days, then passed date
+    if start_period is arg_not_supplied:
+        if calendar_days_back is arg_not_supplied:
+            ## no period or calendar days, use passed date
+            if start_date is arg_not_supplied:
+                raise Exception("Have to specify one of calendar days back, start period or start date!")
+            else:
+                ## Use passed start date
+                pass
+        else:
+            ## Calendar days
+            start_date = n_days_ago(calendar_days_back, end_date)
+    else:
+        ## have a period
+        start_date = get_date_from_period_and_end_date(start_period, end_date)
+
+    return start_date, end_date
+