+++ sysdata/production/capital.py
@@ -3,6 +3,7 @@ import datetime
 import pandas as pd
 
 from syscore.objects import arg_not_supplied, failure, missing_data
+from syscore.pdutils import uniquets
 from sysdata.production.timed_storage import (
     listOfEntriesData,
 )
@@ -52,21 +53,22 @@ class capitalData(listOfEntriesData):
     def _data_class_name(self) -> str:
         return "sysdata.production.capital.capitalForStrategy"
 
-    def get_total_capital_pd_df(self) -> pd.DataFrame:
-        return self.get_capital_pd_df_for_strategy(GLOBAL_STRATEGY)
+    def get_total_capital_pd_series(self) -> pd.Series:
+        return self.get_capital_pd_series_for_strategy(GLOBAL_STRATEGY)
 
-    def get_broker_account_value_pd_df(self) -> pd.DataFrame:
-        return self.get_capital_pd_df_for_strategy(BROKER_ACCOUNT_VALUE)
+    def get_broker_account_value_pd_series(self) -> pd.Series:
+        return self.get_capital_pd_series_for_strategy(BROKER_ACCOUNT_VALUE)
 
-    def get_maximum_account_value_pd_df(self) -> pd.DataFrame:
-        return self.get_capital_pd_df_for_strategy(MAXIMUM_ACCOUNT_VALUE)
+    def get_maximum_account_value_pd_series(self) -> pd.Series:
+        return self.get_capital_pd_series_for_strategy(MAXIMUM_ACCOUNT_VALUE)
 
-    def get_profit_and_loss_account_pd_df(self) -> pd.DataFrame:
-        return self.get_capital_pd_df_for_strategy(ACC_PROFIT_VALUES)
+    def get_profit_and_loss_account_pd_series(self) -> pd.Series:
+        return self.get_capital_pd_series_for_strategy(ACC_PROFIT_VALUES)
 
-    def get_capital_pd_df_for_strategy(self, strategy_name: str) -> pd.DataFrame:
+    def get_capital_pd_series_for_strategy(self, strategy_name: str) -> pd.Series:
         capital_series = self.get_capital_series_for_strategy(strategy_name)
         pd_series = capital_series.as_pd_df()
+        pd_series = uniquets(pd_series).squeeze()
         return pd_series
 
     def get_capital_series_for_strategy(self, strategy_name: str) -> capitalForStrategy:
@@ -115,12 +117,18 @@ class capitalData(listOfEntriesData):
         return current_capital_entry
 
     def update_broker_account_value(
-        self, new_capital_value: float, date: datetime.datetime = arg_not_supplied
+        self, new_capital_value: float,
+            date: datetime.datetime = arg_not_supplied,
+
     ):
+        ## Update account value but also propogate
+        if date is arg_not_supplied:
+            date = datetime.datetime.now()
 
         self.update_capital_value_for_strategy(
             BROKER_ACCOUNT_VALUE, new_capital_value, date=date
         )
+
 
     def update_profit_and_loss_account(
         self, new_capital_value: float, date: datetime.datetime = arg_not_supplied
@@ -314,17 +322,18 @@ class totalCapitalCalculationData(object):
     def get_current_total_capital(self):
         return self.capital_data.get_current_total_capital()
 
-    def get_total_capital(self) -> pd.DataFrame:
-        return self.capital_data.get_total_capital_pd_df()
+    def get_total_capital(self) -> pd.Series:
+        return self.capital_data.get_total_capital_pd_series()
 
-    def get_profit_and_loss_account(self) -> pd.DataFrame():
-        return self.capital_data.get_profit_and_loss_account_pd_df()
 
-    def get_broker_account(self) -> pd.DataFrame:
-        return self.capital_data.get_broker_account_value_pd_df()
+    def get_profit_and_loss_account(self) -> pd.Series():
+        return self.capital_data.get_profit_and_loss_account_pd_series()
 
-    def get_maximum_account(self) -> pd.DataFrame:
-        return self.capital_data.get_maximum_account_value_pd_df()
+    def get_broker_account(self) -> pd.Series:
+        return self.capital_data.get_broker_account_value_pd_series()
+
+    def get_maximum_account(self) -> pd.Series:
+        return self.capital_data.get_maximum_account_value_pd_series()
 
     def get_all_capital_calcs(self) -> pd.DataFrame:
         total_capital = self.get_total_capital()
@@ -429,7 +438,7 @@ class totalCapitalCalculationData(object):
             new_broker_account_value, date=date
         )
 
-        self._update_pandl(profit_and_loss, date)
+        self._update_pandl(profit_and_loss, date=date)
 
     def _update_pandl(self, profit_and_loss: float, date: datetime.datetime):
 
@@ -458,7 +467,10 @@ class totalCapitalCalculationData(object):
         broker_account_value = prev_broker_account_value + delta_value
 
         # Update broker account value
-        self.capital_data.update_broker_account_value(broker_account_value)
+        self.modify_account_values(broker_account_value = broker_account_value,
+                                   propagate=True, are_you_sure=True)
+
+
 
     def modify_account_values(
         self,
@@ -468,6 +480,7 @@ class totalCapitalCalculationData(object):
         acc_pandl: float = arg_not_supplied,
         date: datetime.datetime = arg_not_supplied,
         are_you_sure: bool = False,
+        propagate: bool = True
     ):
         """
         Allow any account valuation to be modified
@@ -485,15 +498,41 @@ class totalCapitalCalculationData(object):
             self.capital_data.update_broker_account_value(
                 broker_account_value, date=date
             )
+        elif propagate:
+            self.propagate_broker_account(date)
 
         if total_capital is not arg_not_supplied:
             self.capital_data.update_total_capital(total_capital, date=date)
+        elif propagate:
+            self.propagate_total_capital(date)
 
         if maximum_capital is not arg_not_supplied:
             self.capital_data.update_maximum_capital(maximum_capital, date=date)
+        elif propagate:
+            self.propagate_maximum_account_value(date)
 
         if acc_pandl is not arg_not_supplied:
             self.capital_data.update_profit_and_loss_account(acc_pandl, date=date)
+        elif propagate:
+            self.propagate_current_pandl(date)
+
+    def propagate_total_capital(self, date):
+        current_total_capital = self.get_current_total_capital()
+        self.capital_data.update_total_capital(current_total_capital, date)
+
+    def propagate_maximum_account_value(self, date):
+        current_max_capital = self.capital_data.get_current_maximum_account_value()
+        self.capital_data.update_maximum_capital(current_max_capital, date)
+
+    def propagate_current_pandl(self, date):
+        current_pandl = self.capital_data.get_current_pandl_account()
+        self.capital_data.update_profit_and_loss_account(current_pandl, date)
+
+    def propagate_broker_account(self, date):
+        broker_account_value = self.capital_data.get_broker_account_value()
+        self.capital_data.update_broker_account_value(
+            broker_account_value, date=date
+        )
 
     def create_initial_capital(
         self,