+++ sysproduction/data/broker.py
@@ -17,7 +17,7 @@ from syscore.objects import (
     missing_contract,
     missing_data
 )
-from syscore.dateutils import Frequency, listOfOpeningTimes
+from syscore.dateutils import Frequency, DAILY_PRICE_FREQ, listOfOpeningTimes
 
 from sysdata.data_blob import dataBlob
 from sysdata.tools.cleaner import apply_price_cleaning
@@ -48,7 +48,7 @@ class dataBroker(productionDataLayerGeneric):
         # Add a list of broker specific classes that will be aliased as self.data.broker_fx_prices,
         # self.data.broker_futures_contract_price ... and so on
 
-        broker_class_list = get_broker_class_list()
+        broker_class_list = get_broker_class_list(data)
         data.add_class_list(broker_class_list)
         return data
 
@@ -122,8 +122,10 @@ class dataBroker(productionDataLayerGeneric):
         broker_prices_raw = \
                 self.get_prices_at_frequency_for_contract_object(contract_object=contract_object,
                                                          frequency = frequency)
+        daily_data = frequency is DAILY_PRICE_FREQ
 
         broker_prices = apply_price_cleaning(data = self.data,
+                                             daily_data=daily_data,
                                              broker_prices_raw = broker_prices_raw,
                                              cleaning_config = cleaning_config)