+++ sysproduction/interactive_diagnostics.py
@@ -1,9 +1,10 @@
-from syscore.dateutils import get_datetime_input, SECONDS_PER_HOUR, openingTimes, listOfOpeningTimes
+from syscore.dateutils import SECONDS_PER_HOUR, openingTimes, listOfOpeningTimes
 from syscore.interactive import (
     get_and_convert,
     run_interactive_menu,
     print_menu_of_values_and_get_response,
     print_menu_and_get_response,
+    true_if_answer_is_yes, get_report_dates
 )
 from syscore.genutils import progressBar
 from syscore.pdutils import set_pd_print_options
@@ -52,7 +53,9 @@ from sysproduction.reporting.report_configs import (
     instrument_risk_report_config,
     min_capital_report_config,
     duplicate_market_report_config,
-    remove_markets_report_config
+    remove_markets_report_config,
+market_monitor_report_config,
+account_curve_report_config
 
 )
 
@@ -134,6 +137,8 @@ nested_menu_of_options = {
         71: "Minimum capital required",
         72: "Duplicate markets",
         73: "Remove markets",
+        74: "Market monitor",
+        75: "P&L account curve"
     }
 
 }
@@ -178,10 +183,10 @@ def roll_report(data):
 
 
 def pandl_report(data):
-    start_date, end_date, calendar_days = get_report_dates(data)
+    start_date, end_date = get_report_dates()
     report_config = email_or_print_or_file(daily_pandl_report_config)
     report_config.modify_kwargs(
-        calendar_days_back=calendar_days, start_date=start_date, end_date=end_date
+         start_date=start_date, end_date=end_date
     )
     run_report(report_config, data=data)
 
@@ -192,10 +197,10 @@ def status_report(data):
 
 
 def trade_report(data):
-    start_date, end_date, calendar_days = get_report_dates(data)
+    start_date, end_date = get_report_dates()
     report_config = email_or_print_or_file(trade_report_config)
     report_config.modify_kwargs(
-        calendar_days_back=calendar_days, start_date=start_date, end_date=end_date
+         start_date=start_date, end_date=end_date
     )
     run_report(report_config, data=data)
 
@@ -232,10 +237,10 @@ def cost_report(data):
     run_report(report_config, data=data)
 
 def slippage_report(data):
-    start_date, end_date, calendar_days = get_report_dates(data)
+    start_date, end_date= get_report_dates()
     report_config = email_or_print_or_file(slippage_report_config)
     report_config.modify_kwargs(
-        calendar_days_back=calendar_days, start_date=start_date, end_date=end_date
+         start_date=start_date, end_date=end_date
     )
     run_report(report_config, data=data)
 
@@ -262,6 +267,36 @@ def remove_markets_report(data):
     run_report(report_config, data=data)
 
 
+def market_monitor_report(data):
+
+    run_full_report = true_if_answer_is_yes('Run normal full report? (alternative is customise dates)')
+    if run_full_report:
+        start_date = arg_not_supplied
+        end_date = arg_not_supplied
+    else:
+        start_date, end_date = get_report_dates()
+
+    report_config = email_or_print_or_file(market_monitor_report_config)
+    report_config.modify_kwargs(
+         start_date=start_date, end_date=end_date
+    )
+    run_report(report_config, data = data)
+
+def account_curve_report(data: dataBlob):
+    run_full_report = true_if_answer_is_yes('Run normal full report? (alternative is customise dates)')
+    if run_full_report:
+        start_date = arg_not_supplied
+        end_date = arg_not_supplied
+    else:
+        start_date, end_date = get_report_dates()
+
+    report_config = email_or_print_or_file(account_curve_report_config)
+    report_config.modify_kwargs(
+         start_date=start_date, end_date=end_date
+    )
+    run_report(report_config, data = data)
+
+
 def email_or_print_or_file(report_config):
     ans = get_and_convert(
         "1: Print or 2: email or 3: file or 4: email and file?",
@@ -280,27 +315,6 @@ def email_or_print_or_file(report_config):
         report_config = report_config.new_config_with_modified_output("emailfile")
 
     return report_config
-
-
-def get_report_dates(data):
-    end_date = get_datetime_input("End date for report?\n", allow_default=True)
-    start_date = get_datetime_input(
-        "Start date for report? (SPACE to use an offset from end date)\n",
-        allow_no_arg=True,
-    )
-    if start_date is None:
-        start_date = arg_not_supplied
-        calendar_days = get_and_convert(
-            "Calendar days back from %s?" % str(end_date),
-            type_expected=int,
-            allow_default=True,
-            default_value=1,
-        )
-
-    else:
-        calendar_days = arg_not_supplied
-
-    return start_date, end_date, calendar_days
 
 
 # logs emails errors
@@ -543,8 +557,7 @@ def get_order_pd(
     list_method="get_historic_instrument_order_ids_in_date_range",
     getter_method="get_historic_instrument_order_from_order_id",
 ):
-    start_date = get_datetime_input("Start Date", allow_default=True)
-    end_date = get_datetime_input("End Date", allow_default=True)
+    start_date, end_date = get_report_dates()
 
     data_orders = dataOrders(data)
     list_func = getattr(data_orders, list_method)
@@ -774,6 +787,8 @@ dict_of_functions = {
     71: min_capital_report,
     72: duplicate_market_report,
     73: remove_markets_report,
+    74: market_monitor_report,
+    75: account_curve_report
 
 }